chore: move test asl to new folder
