+++ infrastructure/product/stream_processor_testing/store_intercepted_events.asl
